Handler for POST method in a non-CGI situation
- Uploads file if request is multipart/form-data (only deals with first element of form)
- Client max body size now is setted from config file
- Hard limit of 4Mb for request body (closes connection if threshold is breached)
- Soft limit of 1Mb max. Reads request body and ignores it, returning 413 to client.

+++ src/VirtualServer.cpp
@@ -165,7 +165,7 @@ void VirtualServer::setBodySize(std::string& directiveValue)
                                directiveValue + "'");
         throw std::runtime_error("");
     }
-    _clientMaxBodySize = static_cast<uint16_t>(tmpBodySize);
+    _clientMaxBodySize = static_cast<size_t>(tmpBodySize);
 }
 
 void VirtualServer::setLocation(std::pair<std::string, Location>& location)
@@ -179,7 +179,7 @@ uint32_t VirtualServer::getHost(void) const { return _host; }
 
 uint16_t VirtualServer::getPort(void) const { return _port; }
 
-int VirtualServer::getBodySize(void) const { return _clientMaxBodySize; }
+size_t VirtualServer::getBodySize(void) const { return _clientMaxBodySize; }
 
 Location* VirtualServer::getLocation(std::string resource)
 {
